+++ concordium-sdk/src/main/java/com/concordium/sdk/crypto/elgamal/ElgamalPublicKey.java
@@ -9,6 +9,8 @@ import lombok.ToString;
 import org.apache.commons.codec.DecoderException;
 import org.apache.commons.codec.binary.Hex;
 
+import java.util.Arrays;
+
 @RequiredArgsConstructor
 @Getter
 @ToString
@@ -23,12 +25,22 @@ public class ElgamalPublicKey {
     @JsonCreator
     public static ElgamalPublicKey from(String hexKey) {
         try {
-            return new ElgamalPublicKey(Hex.decodeHex(hexKey));
+            return from(Hex.decodeHex(hexKey));
         } catch (DecoderException e) {
             throw new IllegalArgumentException("Cannot create ElgamalPublicKey", e);
         }
     }
 
+    /**
+     * Creates an Elgamal Public Key from {@link byte} Array.
+     *
+     * @param bytes Input Byte Array
+     * @return Instance of {@link ElgamalPublicKey}
+     */
+    public static ElgamalPublicKey from(final byte[] bytes) {
+        return new ElgamalPublicKey(Arrays.copyOf(bytes, bytes.length));
+    }
+    
     @Override
     @JsonValue
     public String toString() {