+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/blocksummary/updates/queues/Description.java
@@ -2,6 +2,7 @@ package com.concordium.sdk.responses.blocksummary.updates.queues;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
@@ -18,6 +19,7 @@ public class Description {
     private final String description;
 
     @JsonCreator
+    @Builder
     public Description(@JsonProperty("name") String name, @JsonProperty("url") String url, @JsonProperty("description") String description) {
         this.name = name;
         this.url = url;